Create task.ps1
This powershell script is used to deploy SQL scripts in a server.

There is an in house built runasuser.exe utility where the credentials of the account used to execute (which has access on DB).

Powershell script executes SQL script susing sqlcmd.exe. The credentials to use are fetched from the utility runasuser.exe.